+++ src/main/kotlin/no/deltahouse/kotlinbeer/model/domain/Transaction.kt
@@ -8,6 +8,6 @@ class Transaction(transaction: TransactionDAO) {
     val id: Long = transaction.id
     val user: UserDAO = transaction.user
     val previousBalance: Int = transaction.previousBalance
-    val balanceChange: Int = transaction.balanceChange
+    val balanceChange: Short = transaction.balanceChange
     val transactionDate: ZonedDateTime = transaction.transactionDate
 }
